Association/requirements.txt
absl-py==0.6.1
astor==0.7.1
gast==0.2.0
grpcio==1.16.0
Markdown==3.0.1
numpy==1.14.5
protobuf==3.6.1
scipy==1.1.0
six==1.11.0
tensorboard==1.10.0
tensorflow-gpu==1.10.0
termcolor==1.1.0
Werkzeug==0.14.1
